+++ src/views/blogs/metadata.blade.php
@@ -1,0 +1,8 @@
+@extends("dicms::layouts.admin.index", ['template' => $template])
+
+@section('index_content')
+    <div class="alert alert-warning">
+        {{ __('dicms-blog::blogger.metadata.warning') }}
+    </div>
+    <livewire:metadata-editor :container="$obj" />
+@endsection